Logic to build/test deb package for master
The main change in this commit is to give the build script the ability
to figure out the version string based on git info: the main difference
from the make versioning is that we use Debian style master versions,
being 0.0~[iso-date].[git-hash]-1.

So now by default the Pipeline will build and test only the master code.
To trigger a build/test of a tagged Mender version, we need to manually
trigger a Pipeline with MENDER_VERSION=X.X.X

See also a fix in the test itself to verify "mender -version" otput for
the master version testing.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>