Enable systemd after successful install and configure of mender-client
Copy support/mender.service to debian/mender-client.service and let
debhelper do the rest of the magic.

Note the change in the name: the systemd service shall have the same
name as the package. We had to change this soon or later as an step
towards unifying with Debian upstream recipe.

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>