Add new test for the interactive installation of the client package
Refactor the current code into a base class so that both cases
(interactive and non interactive) can share all the device checks.

Improved robustness of how to check the full cycle of the mender client
daemon with a simple loop and timeout.

Added logic to check mender process is running/stopped on systemd
related checks and for the contents of mender.conf file.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>

squash-there