Do not allow acceptance test failures anymore
This was a temporary solution to be able to publish 2.1.x mender client
packages, but allowing them to fail for so long has made the master (and
therefore 2.2.x) tests to break.

Until we get QA-123, if we need to publish a new deb package for an old
mender client (2.1.x) we can use this new SKIP_TESTS build variable.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>