MEN-4092: Build mender-shell debian package
Add recipe to build mender-shell package the same way we build
mender-client one.

Minor fixes on mender-client recipe to keep them aligned, notably
setting Section to admin.

Rename jobs of the pipeline to make it explicit that only mender-client
is published in the legacy path, mender-shell will only be published in
the new repository.

Acceptance tests fix to not override mender-client-deb-version, instead
output the version string into [package]-deb-version, so that two
separate files are created for the two packages.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>