[pipeline] Add lock around APT repository update
When two pipelines run in parallel (which happens on releases) there is
a high change of corrupting the repository.

Prevent it by locking the S3 bucket that we are about to update.

Note that this is not atomic and there is a chance of forever block one
of the pipelines, in which case it will be killed by GitLab timeout. Not
a very high risk.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>