[test_install_mender_sh.py] De-scope some test cases
Remove the channel parametrization for test_client and test_connect,
reducing the overall test cases for this file from 9 to 5.

The channel switch is already tested in the main test case, and it is
very low risk to remove it from the others.

Currently the full test suite takes more than 1 hour, exceeding the
timeout in CI.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>