Work around when master is exactly on a tag name.
If we have requested to build master in the build parameters, we want
a SHA-based package version, even if it matches the tag exactly. This
is also expected by our tests, which fail without this fix. And it's
also important so that we don't upload the wrong package.

However, the reported version by the tool should be the tag name, if
it's possible. So testing for this when building master is wrong,
because it can be both a tag name and a SHA, depending on what is
built.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>