MEN-4715: Move the script to use /etc/apt/sources.list.d/mender.list
This changes the install script to start using
'/etc/apt/sources.list.d/mender.list' as the source file from which mender is
installed.

As an additional security measure, it searches '/etc/apt/sources.list' for an
existing mender configuration, and removes it if present, so that the new
sources will be respected.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>