dpkg-buildpackage: Stop using --build=binary and let the caller specify
According to its man page, --build=binary is alias for --build=any,all
which resulted in the three mender-client_..._<arch>.changes files
containing the checksum for mender-client-dev and eventually making
reprepro fail when trying to upload the same package (-dev) twice!

Use instead "any" on the for loop for the architectures, and explicitly
pass "all" for the platform independent packages. This creates a
dedicated change files (..._all.changes) for the mender-client with only
the mender-client-dev checksum.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>