Build arch. dependent and independent packages together.
This is faster, and avoids a problem with old recipes that do not have
architecture independent packages. Asking for them explicitly produces
a failure, but asking for both at the same time just silently omits
them, which is what we want.

The only downside is that they are built once for each architecture,
but this has a negligible performance impact, and no practical effect.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>