test_install_mender_sh.py: Run tests on x86-64 container
The purpose of these tests are on verifying the different options of the
install script... and for that we don't need QEMU or ARM emulation.

With this change, we migrate all the tests to use a debian:buster
container, and add *one* test for Raspberry OS.

Locally the tests run now in 10 minutes: 3 for the converted tests and 7
for the new Raspberry OS test.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>