pipeline: publish: Remove legacy "latest" copy and rename docs publish
Job publish:s3:legacy:mender-client:latest would make a copy for old
versions of mender-convert to install the client. Last version to use
this method was mender-convert 2.2.x, now EOL

On the other hand, the job lish:s3:legacy:mender-client had a misleading
name, as the published package is still in use for direct download from
the Mender Docs site, so rename it to be explicit.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>