pipeline: apt-repo: Do not override existing packages
It could happen that a given release re-uses an old component already
tagged and published and we need to skip copying it. Otherwise we are
corrupting the APT repository indexes.

In any case shall we replace a package, the manual copy is meant only to
workaround the publishing of bugfix releases, where reprepro refuses to
publish it for us due to the versioning scheme.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>