pipeline: Publish and reuse Docker images
To always build the Docker images from scratch is just a waste... This
commit reworks the pipeline so that the images are only built on changes
to the relevant files; and the packages are build either pulling the
images from upstream or loading the local images.

The docker-build-package script has been reworked to take the package to
build as an input, with the intention in the future to select via
pipeline variables which packages to build.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>