MEN-5410: Build packages for Debian and Ubuntu two latest releases
Our policy of building from "oldstable" and support two releases this
way has got now into trouble with the upgrade from libffi6 to libffi7 in
latest Debian bullseye.

This commit reworks many parts of the build process to build and publish
deb packages for the last two releases of Debian (Buster and Bullseye)
and the last two of Ubuntu (Bionic and Focal).

Notable changes:
* Dockerfile: adapt to work on the four base imgages. Note that Ubuntu
  does not support foreign architectures in their main repository and a
  ports.ubuntu.com needs to be added. Note also the special handling for
  libffi6/libffi7
* docker-build- wrapper scripts: Modify to build and use the 2 distros x
  2 releases x 3 architectures = 12 builder images
* .gitlab-ci.yml: Adapt to the above accordingly
* recipes and mender-deb-package script: Move away from the hardcoded
  Depends (which brought the breaking problem in the first place) and
  use instead Debian tools to figure out the dependencies. For the case
  where we use our custom toolchain to support ARMv6 (this is, Debian
  for arm 32 bits) post-process the recipes to get the hardcoded
  pre-generated dependencies.

Some minor clean-ups in the recipes were also done.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>