MEN-5410: Append distribution and release to the package version
The original approach had the limitation that the produced deb packages
will have the same version (and name) on different platforms, which will
error to update the repository (all distributions in our repo share the
same pool).

Removing the tentative directory hierarchy also helped with the unknown
on publishing, the acceptance tests and reduces risk on the changes to
be made for the reprepro wrapper script.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>