perf(mender-deb-package): Either build orig tarball or deb package
Previously a job requiring only the orig tarball would anyway compile
and package the debian package. This takes a considerable amount of time
and the result is thrown away by CI.

The benefit was that if `dh_make` failed to create the tarball then the
subsequent `dpkg-buildpackage` would fail in the same job, while now it
would silently pass and fail in the following jobs.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>