ci: Better split orig and deb packages
Although the motivation for this change is to workaround a limitation in
our pipelines (iow by using the `parallel` matrix all the jobs will get
the artifacts from previous stages, which means that the orig packages
will be passed around), I believe that this split makes sense and keeps
the design cleaner.

Basically we'll have a dedicated directory `orig` for the job that
produces such files, and the builds will consume them from there
(instead of making an ad hoc mv for each build) and save their artifcts
in a different directory (split per job, as before).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>