ci: Remove explicit dependencies for publish jobs
By default, GitLab will load all artifacts from previous stages, which
means that we will also load the orig packages that we don't intend to
publish.

However this is an small price to pay in order to not hard-code all the
dependencies that are auto-generated with the parallel matrix.

The previous commit moved the orig packages out of `output/opensource`
so the publish job(s) themselves will just ignore them.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>