build: add the ability to override the image version used
This adds the possibility to override the image version for each of our build
containers, for instance if you want to test this along with an image built from
a PR in `mender-test-containers`.

Then simply override the `IMAGE_VERSION` variable with the sha from your
container output:

Something like:

```
18 pushing registry.gitlab.com/northern.tech/mender/mender-test-containers:mender-dist-packages-builder-debian-bullseye-amd64-987774732
```

Would mean: `IMAGE_VERSION=987774732`.

The default is still `master`.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>