feat: Move post install script for setup to `mender-setup`
Future releases of `mender-client` won't support the `setup` command and
this script needs to be called its own package instead.

Note that the `postinst` script has been moved to `mender-setup` package
while the `postrm` script remains in `mender-client`. The reasoning is
that only `mender-client` should own its configuration file, even if it
is (or it can be) created with a separate tool.

Ticket: MEN-6765
Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>