fix: Enable/disable systemd services with postinst/postrm scripts
The `dh` tools need to know about the `.service` files in order to add
logic to post install and post remove scripts to enable and disable the
services.

Instead of just manually copying the `.service` file to the `debian`
directory as we do for the other recipes, let's use the recommended
setup from Debian's systemd team:
* https://wiki.debian.org/Teams/pkg-systemd/Packaging

Modify also the test to use `systemctl` commands rather than manually
checking the configuration paths, which is more error prone.

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>