fix: Handle diversion of files when upgrading to `mender-client` 4.0
Based on the research from Kristian Amlie. See:
* https://github.com/kacf/mender-dist-packages/commit/0dac3e66097aa54fdb071ea3492a14c8840654dc

With the new packaging, the files that used to belong to `mender-client`
are now owned by `mender-auth` and `mender-update`.

Add `preinst` scripts to the new packages to divert the files and a
`postinst` to `mender-client` to restore the diversions. See inline
comments for more context.

Added a test to verify the upgrade path.

Changelog: None
Ticket: MEN-6945

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>