test: Improve the method to verify installed and purged packages
The motivation to do this _now_ is that with the epoch on the
`mender-client` package, the version technically installed (`1:..`) is
different that the version string in the filename.

To rely on the stdout of a 3rd party tool is fragile and doesn't add any
test value, so move all the tests to use the existing `dpkg --status`
method.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>