feat: add support for mender-client4, mender-auth and mender-update components
From now on, mender-client4 will be a default component in place of
mender-client; however, the script falls back to mender-client (legacy,
go client) for Ubuntu jammy and focal and for Debian bullseye and
buster. As we don't support just yet other distributions, it means that
the only way to install the Mender Client version 4 is to specify its
components explicitly from the command line. In the future, when we'll
support other distributions (e.g., bookworm), mender-client4 will be the
default.

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>