revert: "test: Remove explicit install of `mender-update`"
This reverts commit d3f5a5991391fff1f067ad2c4cc1d72229eb2110.

We cannot remove the explicit install of `mender-update` because the test
will fail with as follows if the add-on packages (e.g., `mender-connect`)
are not upgraded at the same time when upgrading from `mender-client` to
`mender-client4`:

```
The following packages have unmet dependencies:
 mender-client4 : Conflicts: mender-client but 1:3.5.2-1+debian+buster is to be installed
E: Unable to correct problems, you have held broken packages.
```

This happens on release pipelines, where we build `mender-connect` and
`mender-configure` using an old tag, therefore not triggering an upgrade
at the same time we install `mender-client4` to upgrade the Mender client.

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>