chore: Made docker-build-package and mender-deb-package compatible with arm image
Added platform to the docker run command that specifies armv6 if
architecture is armhf.

Updated the mender-deb-package to use the same build process for armhf as for
native builds (using an image emulating armhf). See following PR:
https://github.com/mendersoftware/mender-test-containers/pull/514

Note: set the variable IMAGE_VERSION and the image tag to
"feature-armhf-native". This should be changed to "master" if the
branch gets merged to master.

Ticket: MEN-6875

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>