chore: Better document and make robust the cmake version workaround
The main change is to remove the whole line rather than modify the
specific version, so that we can ignore changes in the original version.

Along the way: add an explicit check for the package name, move from
checking the file to grepping for the string to delete, print the diff,
and add a block comment to justify all this hacking.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>