fix: Add missing `dbus` dependency for `mender-auth`
And for the add-ons that did not have it.

The regular Debian distro, and presumably most of the forks, do come
with `dbus` installed so this went unnoticed for very long. The issue
shows on official variant Debian minbase.

The add-ons depend on `mender-auth` so indirectly `dbus` would be
installed. However as the add-ons explicitly use DBus is correct to have
it in their dependencies too.

Ticket: MEN-7742

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>