test: mark tests as flaky
This is a temporary workaround because we will be moving away from this
special virtual device later this year. This shouldn't have any real
effect on the packages we tests, as the failure to boot is unrelated.
There have been previous attempts of fixing this by increasing the timeout
to 15 minutes; however, it's still hanging on boot. This is a best-effort
fix - as we didn't see the value in debugging a virtual device we know we
will move away from.

Ticket: QA-975

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>