chore: Add an architecture check to armhf packages
`armhf` can mean both ARMv6 and ARMv7. We build our `armhf`
packages for ARMv7 so let's make sure they are not installed on
ARMv6 devices.

The easiest and most user-friendly way to perform such a check is
in the packages' `preinst` scriptlets that run early in the
installation process.

Ticket: QA-1085
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>